requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dateutil>=2.8.2
tqdm>=4.66.0
openai>=1.0.0
anthropic>=0.7.0
python-dotenv>=1.0.0